requirements.txt
pyroblack
tgcrypto
uvicorn
fastapi[all]
python-dotenv
aiofiles
aiohttp